Conflicts:
	CPSC585/CPSC585/src/GameEngine.cpp
	CPSC585/CPSC585/src/Menu.cpp

+++ CPSC585/CPSC585/src/AnalogEvent.cpp
@@ -1,4 +1,4 @@
 #include "AnalogEvent.h"
 
-AnalogEvent::AnalogEvent(int xval, double normXVal, int yval, double normYVal, bool clickedIn) 
+AnalogEvent::AnalogEvent(int xval, float normXVal, int yval, float normYVal, bool clickedIn) 
 		: XValue(xval), normXValue(normXVal), YValue(yval), normYValue(normYVal), clicked(clickedIn){}
