+++ CPSC585/CPSC585/src/RightAnalogEvent.h
@@ -5,7 +5,7 @@
 class RightAnalogEvent : public AnalogEvent {
 public:
 	RightAnalogEvent(){};
-	RightAnalogEvent(int xval,double normXVal, int yval,double normYVal, bool clickedIn):AnalogEvent(xval, normXVal, yval, normYVal, clickedIn){};
+	RightAnalogEvent(int xval,float normXVal, int yval,float normYVal, bool clickedIn):AnalogEvent(xval, normXVal, yval, normYVal, clickedIn){};
 	EventTypes::EventType RightAnalogEvent::getType() const{
 		return EventTypes::RIGHT_ANALOG;
 	}
